+++ examples/multi_roi_example.py
@@ -1,0 +1,32 @@
+import logging
+
+import numpy as np
+from matplotlib import pyplot as plt
+
+from roipoly import MultiRoi
+
+logging.basicConfig(format='%(levelname)s ''%(processName)-10s : %(asctime)s '
+                           '%(module)s.%(funcName)s:%(lineno)s %(message)s',
+                    level=logging.INFO)
+
+# Create image
+img = np.ones((100, 100)) * range(0, 100)
+
+# Show the image
+fig = plt.figure()
+plt.imshow(img, interpolation='nearest', cmap="Greys")
+plt.title("Click on the button to add a new ROI")
+
+# Draw multiple ROIs
+multiroi_named = MultiRoi(roi_names=['My first ROI', 'My second ROI'])
+
+# Draw all ROIs
+plt.imshow(img, interpolation='nearest', cmap="Greys")
+roi_names = []
+for name, roi in multiroi_named.rois.items():
+    roi.display_roi()
+    roi.display_mean(img)
+    roi_names.append(name)
+plt.legend(roi_names, bbox_to_anchor=(1.2, 1.05))
+plt.show()
+