+++ examples/old_example.py
@@ -1,4 +1,4 @@
-"""Test if the old example.py still works
+"""Test if the old example still works
 (commit 80737fa9d66d839e52b40d7cadcf02e143a86b4a)"""
 import pylab as pl
 from roipoly import roipoly
@@ -12,7 +12,7 @@ pl.colorbar()
 pl.title("left click: line segment         right click: close region")
 
 # let user draw first ROI
-ROI1 = roipoly(roicolor='r') #let user draw first ROI
+ROI1 = roipoly(roicolor='r')  # let user draw first ROI
 
 # show the image with the first ROI
 pl.imshow(img, interpolation='nearest', cmap="Greys")
@@ -21,7 +21,7 @@ ROI1.displayROI()
 pl.title('draw second ROI')
 
 # let user draw second ROI
-ROI2 = roipoly(roicolor='b') #let user draw ROI
+ROI2 = roipoly(roicolor='b')  # let user draw ROI
 
 # show the image with both ROIs and their mean values
 pl.imshow(img, interpolation='nearest', cmap="Greys")