+++ roipoly/__init__.py
@@ -1,4 +1,4 @@
-from .roipoly import RoiPoly  # noqa
+from .roipoly import RoiPoly, MultiRoi  # noqa
 from .roipoly import roipoly  # noqa / for compatibility with old versions
 
 from .version import __version__  # noqa