+++ roipoly/roipoly.py
@@ -1,11 +1,5 @@
 """Draw polygon regions of interest (ROIs) in matplotlib images,
 similar to Matlab's roipoly function.
-
-See the file example.py for an application.
-
-Created by Joerg Doepfert 2014 based on code posted by Daniel
-Kornhauser.
-
 """
 
 import sys
@@ -15,6 +9,7 @@ import warnings
 import numpy as np
 import matplotlib.pyplot as plt
 from matplotlib.path import Path as MplPath
+from matplotlib.widgets import Button
 
 logger = logging.getLogger(__name__)
 
@@ -27,8 +22,23 @@ def deprecation(message):
 
 class RoiPoly:
 
-    def __init__(self, fig=None, ax=None, color='b',
-                 roicolor=None, show_fig=True):
+    def __init__(self, fig=None, color='b',
+                 roicolor=None, show_fig=True, close_fig=True):
+        """
+
+        Parameters
+        ----------
+        fig: matplotlib figure
+            Figure on which to create the ROI
+        color: str
+           Color of the ROI
+        roicolor: str
+            deprecated, use `color` instead
+        show_fig: bool
+            Display the figure upon initializing a RoiPoly object
+        close_fig: bool
+            Close the figure after finishing ROI drawing
+        """
 
         if roicolor is not None:
             deprecation("Use 'color' instead of 'roicolor'!")
@@ -37,18 +47,16 @@ class RoiPoly:
         if fig is None:
             fig = plt.gcf()
 
-        if ax is None:
-            ax = plt.gca()
-
         self.start_point = []
         self.end_point = []
         self.previous_point = []
         self.x = []
         self.y = []
         self.line = None
+        self.completed = False  # Has ROI drawing completed?
         self.color = color
         self.fig = fig
-        self.ax = ax
+        self.close_figure = close_fig
 
         # Mouse event callbacks
         self.__cid1 = self.fig.canvas.mpl_connect(
@@ -64,7 +72,7 @@ class RoiPoly:
         if sys.flags.interactive:
             plt.show(block=False)
         else:
-            plt.show()
+            plt.show(block=True)
 
     def get_mask(self, current_image):
         ny, nx = np.shape(current_image)
@@ -160,10 +168,9 @@ class RoiPoly:
                 ax.add_line(self.line)
                 self.fig.canvas.draw()
                 self.line = None
+                self.completed = True
 
-                if sys.flags.interactive:
-                    pass
-                else:
+                if not sys.flags.interactive and self.close_figure:
                     #  Figure has to be closed so that code can continue
                     plt.close(self.fig)
 
@@ -179,6 +186,80 @@ class RoiPoly:
     def displayROI(self, *args, **kwargs):
         deprecation("Use 'display_roi' instead of 'displayROI'!")
         return self.display_roi(*args, **kwargs)
+
+
+class MultiRoi:
+    def __init__(self,
+                 fig=None, ax=None,
+                 roi_names=None,
+                 color_cycle=('b', 'g', 'r', 'c', 'm', 'y', 'k')
+                 ):
+        """
+
+        Parameters
+        ----------
+        fig: matplotlib figure
+            Figure on which to draw the ROIs
+        ax: matplotlib axes
+           Axes on which to draw the ROIs
+        roi_names: list of str
+            Optional names for the ROIs to draw.
+            The ROIs can later be retrieved by using these names as keys for
+            the `self.rois` dictionary. If None, consecutive numbers are used
+            as ROI names
+        color_cycle: list of str
+            List of matplotlib colors for the ROIs
+        """
+
+        if fig is None:
+            fig = plt.gcf()
+        if ax is None:
+            ax = fig.gca()
+
+        self.color_cycle = color_cycle
+        self.roi_names = roi_names
+        self.fig = fig
+        self.ax = ax
+        self.rois = {}
+
+        self.make_buttons()
+
+    def make_buttons(self):
+        ax_add_btn = plt.axes([0.7, 0.02, 0.1, 0.04])
+        ax_finish_btn = plt.axes([0.81, 0.02, 0.1, 0.04])
+        btn_finish = Button(ax_finish_btn, 'Finish')
+        btn_finish.on_clicked(self.finish)
+        btn_add = Button(ax_add_btn, 'New ROI')
+        btn_add.on_clicked(self.add)
+        plt.show(block=True)
+
+    def add(self, event):
+        """"Add a new ROI"""
+
+        # Only draw a new ROI if the previous one is completed
+        if self.rois:
+            if not all(r.completed for r in self.rois.values()):
+                return
+
+        count = len(self.rois)
+        idx = count % len(self.color_cycle)
+        logger.debug("Creating new ROI {}".format(count))
+        if self.roi_names is not None and idx < len(self.roi_names):
+            roi_name = self.roi_names[idx]
+        else:
+            roi_name = str(count + 1)
+
+        self.ax.set_title("Draw ROI '{}'".format(roi_name))
+        plt.draw()
+        roi = RoiPoly(color=self.color_cycle[idx],
+                      fig=self.fig,
+                      close_fig=False,
+                      show_fig=False)
+        self.rois[roi_name] = roi
+
+    def finish(self, event):
+        logger.debug("Stop ROI drawing")
+        plt.close(self.fig)
 
 
 # For compatibility with old version