+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## [unreleased]
 * remove logger from example
-
+* remove list comprehension from basic example
+* remove unused `self.ax` from `RoiPoly`
+* add docstring to `RoiPoly`
+* Allow drawing of multiple ROIs (`MultiRoi`class)
 ## [0.3.0] - 2019-03-06
 * fix readme
 * add makefile